Delete confirmation + TODO added for confirmation modal

+++ angular/src/app/admin/categories/category-detail/category-detail.component.ts
@@ -64,7 +64,10 @@ export class CategoryDetailComponent implements OnInit {
     }
 
     deleteCategory(): void {
-        this.store$.dispatch(CategoriesActions.deleteCategory({ id: this.form.value.id }));
+        // TODO: Create seperate action to trigger confirmation modal?
+        if (confirm('Are you sure?')) {
+            this.store$.dispatch(CategoriesActions.deleteCategory({ id: this.form.value.id }));
+        }
     }
 
     async photoSelected(files: File[]): Promise<void> {
